move vtk_version_info to core._vtk_core and make version_info available from pv
